Removed the Person/Geography equivalence classes with inconsistent naming style.
