--- TestStack.Dossier/EquivalenceClasses/PersonEquivalenceClasses.cs
@@ -1,123 +1,0 @@
-using TestStack.Dossier.DataSources.Dictionaries;
-using TestStack.Dossier.DataSources.Generators;
-
-// ReSharper disable once CheckNamespace
-namespace TestStack.Dossier
-{
-    /// <summary>
-    /// Extension methods that describe equivalence classes for generating anonymous person-related values.
-    /// </summary>
-    public static class PersonEquivalenceClasses
-    {
-        private static Words _personEmailAddressSource;
-        private static Words _uniquePersonEmailAddressSource;
-
-        /// <summary>
-        /// Generate and return an email address.
-        /// </summary>
-        /// <param name="fixture">The fixture to generate a email for</param>
-        /// <returns>The generated email</returns>
-        public static string EmailAddress(this AnonymousValueFixture fixture)
-        {
-            return fixture.DictionaryFor(FromDictionary.PersonEmailAddress).Next();
-        }
-
-        /// <summary>
-        /// Generate and return a unique email address (within the fixture).
-        /// </summary>
-        /// <param name="fixture">The fixture to generate a unique email for</param>
-        /// <returns>The generated unique email</returns>
-        public static string UniqueEmailAddress(this AnonymousValueFixture fixture)
-        {
-            if (_uniquePersonEmailAddressSource == null)
-            {
-                if (_personEmailAddressSource == null) _personEmailAddressSource = fixture.DictionaryFor(FromDictionary.PersonEmailAddress);
-                var generator = new SequentialGenerator(0, _personEmailAddressSource.Data.Count, listShouldBeUnique: true);
-                _uniquePersonEmailAddressSource = new Words(generator, new CachedFileDictionaryRepository(),  FromDictionary.PersonEmailAddress);
-            }
-
-            return _uniquePersonEmailAddressSource.Next();
-        }
-
-        /// <summary>
-        /// Generate and return a language name.
-        /// </summary>
-        /// <param name="fixture">The fixture to generate a language for</param>
-        /// <returns>The generated language</returns>
-        public static string Language(this AnonymousValueFixture fixture)
-        {
-            return fixture.DictionaryFor(FromDictionary.PersonLanguage).Next();
-        }
-
-        /// <summary>
-        /// Generate and return a female first name.
-        /// </summary>
-        /// <param name="fixture">The fixture to generate a first name for</param>
-        /// <returns>The generated female first name</returns>
-        public static string FemaleFirstName(this AnonymousValueFixture fixture)
-        {
-            return fixture.DictionaryFor(FromDictionary.PersonNameFirstFemale).Next();
-        }
-
-        /// <summary>
-        /// Generate and return a male or female first name.
-        /// </summary>
-        /// <param name="fixture">The fixture to generate a first name for</param>
-        /// <returns>The generated first name</returns>
-        public static string FirstName(this AnonymousValueFixture fixture)
-        {
-            return fixture.DictionaryFor(FromDictionary.PersonNameFirst).Next();
-        }
-
-        /// <summary>
-        /// Generate and return a male or female full name (first and last names).
-        /// </summary>
-        /// <param name="fixture">The fixture to generate a full name for</param>
-        /// <returns>The generated full name</returns>
-        public static string FullName(this AnonymousValueFixture fixture)
-        {
-            return fixture.DictionaryFor(FromDictionary.PersonNameFull).Next();
-        }
-
-        /// <summary>
-        /// Generate and return a last name.
-        /// </summary>
-        /// <param name="fixture">The fixture to generate a last name for</param>
-        /// <returns>The generated last name</returns>
-        public static string LastName(this AnonymousValueFixture fixture)
-        {
-            return fixture.DictionaryFor(FromDictionary.PersonNameLast).Next();
-        }
-
-        /// <summary>
-        /// Generate and return a male first name.
-        /// </summary>
-        /// <param name="fixture">The fixture to generate a male first name for</param>
-        /// <returns>The generated male first name</returns>
-        public static string MaleFirstName(this AnonymousValueFixture fixture)
-        {
-            return fixture.DictionaryFor(FromDictionary.PersonNameFirstMale).Next();
-        }
-
-        /// <summary>
-        /// Generate and return name suffix.
-        /// </summary>
-        /// <param name="fixture">The fixture to generate a suffix for</param>
-        /// <returns>The generated suffix</returns>
-        public static string Suffix(this AnonymousValueFixture fixture)
-        {
-            return fixture.DictionaryFor(FromDictionary.PersonNameSuffix).Next();
-        }
-
-        /// <summary>
-        /// Generate and return a name title.
-        /// </summary>
-        /// <param name="fixture">The fixture to generate a title for</param>
-        /// <returns>The generated title</returns>
-        public static string Title(this AnonymousValueFixture fixture)
-        {
-            return fixture.DictionaryFor(FromDictionary.PersonNameTitle).Next();
-        }
-
-    }
-}
