+++ TestStack.Dossier/Suppliers/DefaultLastNameValueSupplier.cs
@@ -17,7 +17,7 @@ namespace TestStack.Dossier.Suppliers
         /// <inheritdoc />
         public object GenerateAnonymousValue(AnonymousValueFixture any, Type type, string propertyName)
         {
-            return any.LastName();
+            return any.PersonNameLast();
         }
     }
 }
